+++ front/commonitilvalidation.form.php
@@ -33,6 +33,10 @@
  * ---------------------------------------------------------------------
  */
 
+use Glpi\Event;
+use Glpi\Exception\Http\AccessDeniedHttpException;
+use Glpi\Exception\Http\BadRequestHttpException;
+
 /**
  * @since 0.85
  */
@@ -42,15 +46,13 @@
  * @var CommonITILValidation $validation
  */
 
-use Glpi\Event;
-
 Session::checkLoginUser();
 
 if (!($validation instanceof CommonITILValidation)) {
-    Html::displayErrorAndDie('');
+    throw new BadRequestHttpException();
 }
 if (!$validation->canView()) {
-    Html::displayRightError();
+    throw new AccessDeniedHttpException();
 }
 
 $itemtype = $validation::getItilObjectItemType();
@@ -123,4 +125,5 @@ if (isset($_POST["add"])) {
         Html::back();
     }
 }
-Html::displayErrorAndDie('Lost');
+
+throw new BadRequestHttpException();