+++ front/manuallink.form.php
@@ -34,6 +34,7 @@
  */
 
 use Glpi\Event;
+use Glpi\Exception\Http\BadRequestHttpException;
 use Glpi\Http\Response;
 
 Session::checkValidSessionId();
@@ -105,5 +106,5 @@ if (array_key_exists('purge', $_POST) || array_key_exists('delete', $_POST)) {
         'items_id'    => $items_id
     ]);
 } else {
-    Html::displayErrorAndDie('lost');
+    throw new BadRequestHttpException();
 }