+++ front/search.php
@@ -33,6 +33,8 @@
  * ---------------------------------------------------------------------
  */
 
+use Glpi\Exception\Http\AccessDeniedHttpException;
+
 /** @var array $CFG_GLPI */
 global $CFG_GLPI;
 
@@ -40,7 +42,7 @@ Session::checkCentralAccess();
 Html::header(__('Search'), $_SERVER['PHP_SELF']);
 
 if (!$CFG_GLPI['allow_search_global']) {
-    Html::displayRightError();
+    throw new AccessDeniedHttpException();
 }
 if (isset($_GET["globalsearch"])) {
     $searchtext = trim($_GET["globalsearch"]);