+++ src/Glpi/Search/SearchEngine.php
@@ -40,6 +40,7 @@ use CommonITILObject;
 use Glpi\Application\View\TemplateRenderer;
 use Glpi\Asset\Asset_PeripheralAsset;
 use Glpi\Debug\Profiler;
+use Glpi\Exception\Http\AccessDeniedHttpException;
 use Glpi\Features\TreeBrowse;
 use Glpi\Plugin\Hooks;
 use Glpi\Search\Input\QueryBuilder;
@@ -376,14 +377,14 @@ final class SearchEngine
         if (!$CFG_GLPI['allow_search_all']) {
             foreach ($p['criteria'] as $val) {
                 if (isset($val['field']) && $val['field'] == 'all') {
-                    \Html::displayRightError();
+                    throw new AccessDeniedHttpException();
                 }
             }
         }
         if (!$CFG_GLPI['allow_search_view'] && !array_key_exists('globalsearch', $p)) {
             foreach ($p['criteria'] as $val) {
                 if (isset($val['field']) && $val['field'] == 'view') {
-                    \Html::displayRightError();
+                    throw new AccessDeniedHttpException();
                 }
             }
         }